requirements.txt
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
psycopg2
djangorestframework-simplejwt